requirements.txt
httpx==0.27.2
openai==1.40.0
requests==2.31.0
beautifulsoup4==4.12.3
python-dotenv==1.0.0
pydantic>=2.7.0
pdfminer.six==20231228
fastapi>=0.115.0
uvicorn>=0.30.0
supabase==2.8.1
gotrue==2.8.1
playwright>=1.49.0
